refactor: VocabularySection 분리 및 오디오 재생 기능 통합
