+++ frontend/src/App.tsx
@@ -23,6 +23,7 @@ import type { DailyData, VocabularyItem } from "@/types";
 import { AuthService } from "./services/authService";
 import AuthPage from "./AuthPage";
 import CalendarSection from "./CalendarSection";
+import VocabularySection from "./VocabularySection";
 
 export default function EnglishLearningApp() {
   // 초기 로딩 상태 추가
@@ -38,7 +39,6 @@ export default function EnglishLearningApp() {
   const [darkMode, setDarkMode] = useState(false);
   const [completed, setCompleted] = useState(false);
   const [currentMonth, setCurrentMonth] = useState(new Date());
-  const [showMeaning, setShowMeaning] = useState<Record<number, boolean>>({});
   const [hoveredWordId, setHoveredWordId] = useState<string | null>(null);
   const [expandedSentences, setExpandedSentences] = useState<
     Record<number, boolean>
@@ -363,6 +363,7 @@ export default function EnglishLearningApp() {
       alert(`완료 처리에 실패했습니다.\n${error}`);
     }
   };
+
 
   const getShortPos = (pos: string): string => {
     const posMap: Record<string, string> = {
@@ -710,88 +711,11 @@ export default function EnglishLearningApp() {
             darkMode ? "bg-gray-800" : "bg-white"
           } shadow-lg`}
         >
-          <div className="flex justify-between items-center mb-4">
-            <h2 className="text-2xl font-bold">Key Words</h2>
-            <button
-              onClick={() => {
-                const allShown = filteredVocabularies.every((vocab) => {
-                  const idx = vocabularies.indexOf(vocab);
-                  return showMeaning[idx];
-                });
-                const newState: Record<number, boolean> = {};
-                filteredVocabularies.forEach((vocab) => {
-                  const idx = vocabularies.indexOf(vocab);
-                  newState[idx] = !allShown;
-                });
-                setShowMeaning(newState);
-              }}
-              className={`text-xs font-medium px-3 py-1 rounded-full transition-colors ${
-                darkMode
-                  ? "bg-gray-600 hover:bg-gray-500 text-gray-300"
-                  : "bg-gray-200 hover:bg-gray-300 text-gray-700"
-              }`}
-            >
-              {filteredVocabularies.every((vocab) => {
-                const idx = vocabularies.indexOf(vocab);
-                return showMeaning[idx];
-              })
-                ? "뜻 숨기기"
-                : "뜻 보기"}
-            </button>
-          </div>
-
-          <div className="space-y-4">
-            {filteredVocabularies.map((item) => {
-              const idx = vocabularies.indexOf(item);
-              return (
-                <div
-                  key={item.id}
-                  onClick={() =>
-                    setShowMeaning({ ...showMeaning, [idx]: !showMeaning[idx] })
-                  }
-                  className={`p-4 rounded-lg cursor-pointer transition-colors ${
-                    darkMode
-                      ? "bg-gray-700 hover:bg-gray-600"
-                      : "bg-gray-50 hover:bg-gray-100"
-                  }`}
-                >
-                  <div className="flex items-center gap-2 mb-2">
-                    <span className="font-bold text-lg">{item.word}</span>
-                    <span
-                      className={`text-xs px-2 py-1 rounded ${getPosColor(
-                        item.partOfSpeech,
-                      )}`}
-                    >
-                      {getShortPos(item.partOfSpeech)}
-                    </span>
-                  </div>
-
-                  {showMeaning[idx] && (
-                    <div className="mt-2">
-                      <p className="text-sm mb-2">{item.meaning}</p>
-                    </div>
-                  )}
-
-                  {item.exampleSentence && showMeaning[idx] && (
-                    <div
-                      className={`mt-2 p-3 rounded ${
-                        darkMode ? "bg-gray-600" : "bg-white"
-                      }`}
-                    >
-                      <p className="text-sm italic mb-1">
-                        "{item.exampleSentence}"
-                      </p>
-                      {item.exampleTranslation && (
-                        <p className="text-xs text-gray-500">
-                          {item.exampleTranslation}
-                        </p>
-                      )}
-                    </div>
-                  )}
-                </div>
-              );
-            })}
-          </div>
+          <VocabularySection
+            darkMode={darkMode}
+            vocabularies={vocabularies}
+            filteredVocabularies={filteredVocabularies}
+          />
         </div>
 
         <div className="flex gap-4 mb-6">
